pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "boolean-output", email = "wellsfargo.com"},
]
dependencies = [
    "argparse>=1.4.0",
    "flask-restx>=0.5.1",
    "werkzeug==2.1.2",
    "confluence-rest-library>=1.2.2",
    "faker>=13.15.1",
    "beautifulsoup4>=4.11.1",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
